[WIP] runbot: commit in community triggers a build in enterprise
Basic idea:
A flag (rev_dep_build) is added on the repo object.
If the flag is set, it means that we want that repo to be built along
when the dependency_ids repos are built.

Typically, if we want enterprise repo to be built along with each
community commit, we set this flag to True on enterprise repo.

During the checkout of the repo, if such a reverse dependency exists, the
reverse repo is checked out too.